[refactor] moves 'generic' utl/utensils to render because its functions are only used there
