+++ src/jsdependencies.mk
@@ -25,7 +25,7 @@ src/script/render/graphics/entities.js: \
 
 src/script/render/graphics/markermanager.js: \
 	src/script/render/text/arcmappings.js \
-	src/script/utl/utensils.js
+	src/script/render/utl/utensils.js
 
 src/script/render/graphics/renderast.js: \
 	src/script/render/graphics/constants.js \
@@ -59,16 +59,16 @@ src/script/render/graphics/svgelementfactory.js: \
 src/script/render/graphics/svgutensils.js: \
 	src/script/render/graphics/constants.js \
 	src/script/render/graphics/idmanager.js \
-	src/script/utl/utensils.js
+	src/script/render/utl/utensils.js
 
 src/script/render/text/ast2animate.js: \
-	src/script/utl/utensils.js
+	src/script/render/utl/utensils.js
 
 src/script/render/text/ast2dot.js: \
 	src/script/render/text/arcmappings.js \
 	src/script/render/text/flatten.js \
 	src/script/render/text/textutensils.js \
-	src/script/utl/utensils.js
+	src/script/render/utl/utensils.js
 
 src/script/render/text/ast2doxygen.js: \
 	src/script/render/text/arcmappings.js \
@@ -98,7 +98,7 @@ src/script/render/text/flatten.js: \
 	src/script/render/text/arcmappings.js \
 	src/script/render/text/asttransform.js \
 	src/script/render/text/textutensils.js \
-	src/script/utl/utensils.js
+	src/script/render/utl/utensils.js
 
 src/script/ui/interpreter/animator.js: \
 	src/script/render/graphics/renderast.js \
@@ -119,8 +119,7 @@ src/script/ui/interpreter/editor-events.js: \
 	src/lib/codemirror/mode/mscgen/mscgen.js \
 	src/script/ui/interpreter/uistate.js \
 	src/script/ui/utl/gaga.js \
-	src/script/ui/utl/maps.js \
-	src/script/utl/utensils.js
+	src/script/ui/utl/maps.js
 
 src/script/ui/interpreter/general-actions.js: \
 	src/script/ui/interpreter/animator.js \
@@ -200,7 +199,7 @@ src/script/test/cli/t_actions.js: \
 
 src/script/test/cli/t_normalizations.js: \
 	src/script/cli/normalizations.js \
-	src/script/utl/utensils.js
+	src/script/render/utl/utensils.js
 
 src/script/test/cli/t_validations.js: \
 	src/script/cli/validations.js
@@ -258,8 +257,8 @@ src/script/test/render/text/t_ast2xu.js: \
 
 src/script/test/render/text/t_colorize.js: \
 	src/script/render/text/colorize.js \
-	src/script/test/astfixtures.js \
-	src/script/utl/utensils.js
+	src/script/render/utl/utensils.js \
+	src/script/test/astfixtures.js
 
 src/script/test/render/text/t_flatten.js: \
 	src/script/render/text/flatten.js \
@@ -315,12 +314,12 @@ EMBED_JS_SOURCES=src/script/mscgen-inpage.js \
 	src/script/render/text/asttransform.js \
 	src/script/render/text/flatten.js \
 	src/script/render/text/textutensils.js \
+	src/script/render/utl/utensils.js \
 	src/script/ui/embedding/config.js \
 	src/script/ui/embedding/error-rendering.js \
 	src/script/ui/utl/domutl.js \
 	src/script/ui/utl/exporter.js \
-	src/script/ui/utl/paramslikker.js \
-	src/script/utl/utensils.js
+	src/script/ui/utl/paramslikker.js
 # amd dependencies
 INTERPRETER_JS_SOURCES=src/script/mscgen-interpreter.js \
 	src/lib/canvg/StackBlur.js \
@@ -362,6 +361,7 @@ INTERPRETER_JS_SOURCES=src/script/mscgen-interpreter.js \
 	src/script/render/text/colorize.js \
 	src/script/render/text/flatten.js \
 	src/script/render/text/textutensils.js \
+	src/script/render/utl/utensils.js \
 	src/script/ui/interpreter/animator.js \
 	src/script/ui/interpreter/editor-events.js \
 	src/script/ui/interpreter/general-actions.js \
@@ -376,8 +376,7 @@ INTERPRETER_JS_SOURCES=src/script/mscgen-interpreter.js \
 	src/script/ui/utl/gaga.js \
 	src/script/ui/utl/maps.js \
 	src/script/ui/utl/paramslikker.js \
-	src/script/ui/utl/store.js \
-	src/script/utl/utensils.js
+	src/script/ui/utl/store.js
 # cjs dependencies
 CLI_JS_SOURCES=src/script/cli/mscgen.js \
 	src/script/cli/actions.js \
