+++ src/script/render/graphics/markermanager.js
@@ -7,7 +7,7 @@ if ( typeof define !== 'function') {
     var define = require('amdefine')(module);
 }
 
-define(["../../utl/utensils", "../text/arcmappings"], function(_, map) {
+define(["../utl/utensils", "../text/arcmappings"], function(_, map) {
     "use strict";
 
     var KINDS = {
