+++ src/script/render/text/ast2animate.js
@@ -18,7 +18,7 @@ if ( typeof define !== 'function') {
     var define = require('amdefine')(module);
 }
 
-define(["../../utl/utensils"], function(utl) {
+define(["../utl/utensils"], function(utl) {
     "use strict";
 
     var EMPTY_ARC     = [{kind:"|||"}];
