+++ src/script/render/text/ast2dot.js
@@ -12,7 +12,7 @@ if ( typeof define !== 'function') {
     var define = require('amdefine')(module);
 }
 
-define(["./flatten", "./textutensils", "./arcmappings", "../../utl/utensils"],
+define(["./flatten", "./textutensils", "./arcmappings", "../utl/utensils"],
 function(flatten, txt, map, utl) {
     "use strict";
