+++ src/script/render/text/flatten.js
@@ -8,7 +8,7 @@ if ( typeof define !== 'function') {
     var define = require('amdefine')(module);
 }
 
-define(["./asttransform", "./arcmappings", "../../utl/utensils", "./textutensils"],
+define(["./asttransform", "./arcmappings", "../utl/utensils", "./textutensils"],
 /**
  * Defines some functions to simplify a given abstract syntax tree.
  *
